Clean up license language and usage
