+++ oct2py/compat.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+# Copyright (c) oct2py developers.
+# Distributed under the terms of the MIT License.
+
 import sys
 import ctypes
 import os
