+++ oct2py/demo.py
@@ -1,11 +1,8 @@
-"""
-.. module:: demo
-   :synopsis: Play a demo script showing most of the p2oct api features.
+# -*- coding: utf-8 -*-
+# Copyright (c) oct2py developers.
+# Distributed under the terms of the MIT License.
 
-.. moduleauthor:: Steven Silvester <steven.silvester@ieee.org>
-
-"""
-from __future__ import print_function
+from __future__ import print_function, absolute_import
 import time
 from .compat import PY2
