+++ oct2py/tests/test_conversions.py
@@ -3,8 +3,7 @@ import os
 
 import numpy as np
 
-from oct2py import Oct2Py
-from oct2py.utils import Struct
+from oct2py import Oct2Py, Struct
 from oct2py.compat import unicode, long
