+++ oct2py/tests/test_misc.py
@@ -3,7 +3,6 @@ import glob
 import logging
 import os
 import shutil
-import sys
 import tempfile
 
 import pytest
