+++ oct2py/tests/test_numpy.py
@@ -3,7 +3,6 @@ import os
 
 import numpy as np
 
-
 from oct2py import Oct2Py
