+++ oct2py/tests/test_roundtrip.py
@@ -3,8 +3,7 @@ import os
 
 import numpy as np
 
-from oct2py import Oct2Py, Oct2PyError
-from oct2py.utils import Struct
+from oct2py import Oct2Py, Oct2PyError, Struct
 from oct2py.compat import unicode, long
