+++ oct2py/tests/test_usage.py
@@ -8,8 +8,7 @@ from IPython.display import Image, SVG
 import numpy as np
 import pytest
 
-from oct2py import Oct2Py, Oct2PyError
-from oct2py.utils import Struct
+from oct2py import Oct2Py, Oct2PyError, Struct
 from oct2py.compat import StringIO
